requirements.txt
fastapi[standard]==0.128.0
langchain==1.2.0
langgraph==1.0.5
langchain_openai==1.1.6
pydotenv==0.0.7
duckduckgo_search==8.1.1
ddgs==9.10.0
langchain_community==0.4.1
streamlit==1.52.2
langchain-google-genai==4.1.2
aiosqlite==0.22.1
langgraph-checkpoint-sqlite-3.0.1